+++ examples/with-react/src/core/di/container.ts
@@ -1,0 +1,7 @@
+import { container } from 'tsyringe'
+import { LocalTodoRepository } from '../../features/todo/infrastructure/local-todo-repository'
+import { InjectionTokens } from './injection-tokens'
+
+container.registerSingleton(InjectionTokens.TODO_REPOSITORY, LocalTodoRepository)
+
+export { container }