+++ examples/with-react/src/core/di/injection-tokens.ts
@@ -1,0 +1,3 @@
+export class InjectionTokens {
+  static TODO_REPOSITORY = Symbol.for('TODO_REPOSITORY')
+}