+++ packages/arch/src/use-case/execution-options.ts
@@ -1,3 +1,4 @@
 export interface ExecutionOptions {
   inlineError: boolean
+  invalidateCache: boolean
 }